Rewrited solution.
