+++ packages/ckeditor5-ui/tests/manual/tickets/5328/1.html
@@ -232,6 +232,102 @@
 									8
 								</td>
 							</tr>
+							<tr>
+								<td>
+									1
+								</td>
+								<td>
+									2
+								</td>
+							</tr>
+							<tr>
+								<td>
+									3
+								</td>
+								<td>
+									4
+								</td>
+							</tr>
+							<tr>
+								<td>
+									5
+								</td>
+								<td>
+									6
+								</td>
+							</tr>
+							<tr>
+								<td>
+									7
+								</td>
+								<td>
+									8
+								</td>
+							</tr>
+							<tr>
+								<td>
+									1
+								</td>
+								<td>
+									2
+								</td>
+							</tr>
+							<tr>
+								<td>
+									3
+								</td>
+								<td>
+									4
+								</td>
+							</tr>
+							<tr>
+								<td>
+									5
+								</td>
+								<td>
+									6
+								</td>
+							</tr>
+							<tr>
+								<td>
+									7
+								</td>
+								<td>
+									8
+								</td>
+							</tr>
+							<tr>
+								<td>
+									1
+								</td>
+								<td>
+									2
+								</td>
+							</tr>
+							<tr>
+								<td>
+									3
+								</td>
+								<td>
+									4
+								</td>
+							</tr>
+							<tr>
+								<td>
+									5
+								</td>
+								<td>
+									6
+								</td>
+							</tr>
+							<tr>
+								<td>
+									7
+								</td>
+								<td>
+									8
+								</td>
+							</tr>
 						</tbody>
 					</table>
 				</figure>
@@ -322,7 +418,7 @@
 		width: 100%;
 		height: auto;
 		justify-content: center;
-		padding: 40px;
+		padding: 120px 40px;
 	}
 
 	.container-inner > div {
