+++ packages/ckeditor5-widget/src/widgettoolbarrepository.ts
@@ -321,7 +321,8 @@ function getBalloonPositionData( editor: Editor, relatedElement: ViewElement ) {
 			defaultPositions.southArrowNorth,
 			defaultPositions.southArrowNorthWest,
 			defaultPositions.southArrowNorthEast,
-			defaultPositions.viewportStickyNorth
+			defaultPositions.viewportStickyNorth,
+			defaultPositions.viewportHidden
 		]
 	};
 }
